+++ test/emoji/core_test.clj
@@ -7,14 +7,14 @@
     (is (= (emoji/->emoji "smile")
            "😄"))))
 
-(deftest emoji-test
-  (testing "keyword to emoji"
-    (is (= (emoji/->emoji :smile)
-           "😄"))))
-
 (deftest emojify-test
   (testing "Sentence with emojis"
     (is (= (emoji/emojify "Clojure is awesome :thumbsup:")
            "Clojure is awesome 👍"))
     (is (= (emoji/emojify "Sending :e-mail:...")
            "Sending 📧..."))))
+
+(deftest emojify-all-test
+  (testing "Regular sentence with no intention to use emoji"
+    (is (= (emoji/emojify-all "Sending e-mail with a smile")
+           "Sending 📧 with 🅰 😄"))))